DEV: Autoload files for plugin
